An almost working test_spice_editor.py after major refactoring. Missing management of updates and updating of sub-circuits
