+++ spicelib/editor/qsch_editor.py
@@ -27,7 +27,7 @@ import logging
 from .primitives import format_eng
 from .editor_errors import ComponentNotFoundError, ParameterNotFoundError
 from .base_editor import PARAM_REGEX
-from .spice_editor import UNIQUE_SIMULATION_DOT_INSTRUCTIONS
+from .spice_utils import UNIQUE_SIMULATION_DOT_INSTRUCTIONS
 from .base_schematic import (BaseSchematic, SchematicComponent, Point, ERotation, Line, Text, TextTypeEnum,
                              LineStyle, Shape)
 from ..simulators.qspice_simulator import Qspice
